+++ Config/config.php
@@ -53,6 +53,25 @@ return [
         'integrations' => [
             'mautic.integration.smsapi' => [
                 'class' => SmsapiIntegration::class,
+                'arguments' => [
+                    'event_dispatcher',
+                    'mautic.helper.cache_storage',
+                    'doctrine.orm.entity_manager',
+                    'session',
+                    'request_stack',
+                    'router',
+                    'translator',
+                    'logger',
+                    'mautic.helper.encryption',
+                    'mautic.lead.model.lead',
+                    'mautic.lead.model.company',
+                    'mautic.helper.paths',
+                    'mautic.core.model.notification',
+                    'mautic.lead.model.field',
+                    'mautic.plugin.model.integration_entity',
+                    'mautic.lead.model.dnc',
+                    'mautic.sms.smsapi.gateway',
+                ],
                 'tags' => [
                     'mautic.integration',
                     'mautic.config_integration',